Separate animation wrapper into its own component
